pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "combbandits"
version = "0.1.0"
description = "LLM-CUCB-AT: Regret-Bounded Combinatorial Bandits with Unreliable LLM Action Oracles"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Vikram Kakaria", email = "vk5825@princeton.edu"},
]

dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "openai>=1.0",
    "anthropic>=0.20",
    "tiktoken>=0.5",
    "networkx>=3.1",
    "aiohttp>=3.9",
    "aiosqlite>=0.19",
    "click>=8.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "black", "ruff"]
cluster = ["submitit>=1.5"]

[project.scripts]
combbandits = "combbandits.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
